Move tqdm formatting to helpers since used by both Mapping and Custom SDG
